# docs/swagger.yaml
basePath: /api
definitions:
  apperrors.Code:
    enum:
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    - 7
    type: integer
    x-enum-varnames:
    - InternalCode
    - BadRequestCode
    - NotFoundCode
    - AlreadyExistsCode
    - UnauthorizedCode
    - ForbiddenCode
    - VersionMismatchCode
  apperrors.Error:
    properties:
      code:
        $ref: '#/definitions/apperrors.Code'
      incidentID:
        type: string
      stackTrace:
        type: string
    type: object
...
  contracts.PaginatedRequest:
    properties:
      page:
        type: integer
      size:
        type: integer
    type: object
...
  contracts.Star:
    properties:
      bio:
        type: string
      birthDate:
        type: string
      birthPlace:
        type: string
      createdAt:
        type: string
      deathDate:
        type: string
      deletedAt:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      middleName:
        type: string
    type: object
...
  pagination.PaginatedResponse-contracts_Star:
    properties:
      items:
        items:
          $ref: '#/definitions/contracts.Star'
        type: array
      page:
        minimum: 0
        type: integer
      size:
        minimum: 0
        type: integer
      total:
        type: integer
    type: object
...
  stars.CreateStarRequest:
    properties:
      bio:
        type: string
      birthDate:
        type: string
      birthPlace:
        maxLength: 100
        type: string
      deathDate:
        type: string
      firstName:
        maxLength: 50
        minLength: 1
        type: string
      lastName:
        maxLength: 50
        minLength: 1
        type: string
      middleName:
        maxLength: 50
        type: string
    type: object
  stars.Star:
    properties:
      bio:
        type: string
      birthDate:
        type: string
      birthPlace:
        type: string
      createdAt:
        type: string
      deathDate:
        type: string
      deletedAt:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      middleName:
        type: string
    type: object
  stars.UpdateStarRequest:
    properties:
      bio:
        type: string
      birthDate:
        type: string
      birthPlace:
        maxLength: 100
        type: string
      deathDate:
        type: string
      firstName:
        maxLength: 50
        type: string
      lastName:
        maxLength: 50
        type: string
      middleName:
        maxLength: 50
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: david417program@gmail.com
    name: API Creator Support
    url: https://github.com/DavidMovas/Movies-Reviews
  description: This API is used for managing movies reviews.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Movies Reviews API
  version: 1.0.0
paths:
...
  /stars:
    get:
      description: Get stars
      operationId: get-stars
      parameters:
      - description: Request, if request body empty, default values will be used
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/contracts.PaginatedRequest'
      produces:
      - application/json
      responses:
        "200":
          description: PaginatedResponse of Stars, total number of stars, or nil if
            none found
          schema:
            $ref: '#/definitions/pagination.PaginatedResponse-contracts_Star'
        "400":
          description: Invalid request, invalid parameter or missing parameter
          schema:
            $ref: '#/definitions/apperrors.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperrors.Error'
      summary: Get stars
      tags:
      - stars
    post:
      description: Create star
      operationId: create-star
      parameters:
      - description: Request, can have optional fields
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/stars.CreateStarRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Star
          schema:
            $ref: '#/definitions/stars.Star'
        "400":
          description: Invalid request, invalid parameter or missing parameter
          schema:
            $ref: '#/definitions/apperrors.Error'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperrors.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperrors.Error'
      summary: Create star
      tags:
      - stars
  /stars/{starId}:
    delete:
      description: Delete star by id
      operationId: delete-star-by-id
      parameters:
      - description: Star ID
        in: path
        name: starId
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: Invalid request, invalid parameter or missing parameter
          schema:
            $ref: '#/definitions/apperrors.Error'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperrors.Error'
        "404":
          description: Star not found
          schema:
            $ref: '#/definitions/apperrors.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperrors.Error'
      summary: Delete star by id
      tags:
      - stars
    get:
      description: Get star by id
      operationId: get-star-by-id
      parameters:
      - description: Star ID
        in: path
        name: starId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Star
          schema:
            $ref: '#/definitions/stars.Star'
        "400":
          description: Invalid request, invalid parameter or missing parameter
          schema:
            $ref: '#/definitions/apperrors.Error'
        "404":
          description: Star not found
          schema:
            $ref: '#/definitions/apperrors.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperrors.Error'
      summary: Get star by id
      tags:
      - stars
    put:
      description: Update star by id
      operationId: update-star-by-id
      parameters:
      - description: Star ID
        in: path
        name: starId
        required: true
        type: integer
      - description: Request, can have optional fields
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/stars.UpdateStarRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Star
          schema:
            $ref: '#/definitions/stars.Star'
        "400":
          description: Invalid request, invalid parameter or missing parameter
          schema:
            $ref: '#/definitions/apperrors.Error'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperrors.Error'
        "404":
          description: Star not found
          schema:
            $ref: '#/definitions/apperrors.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperrors.Error'
      summary: Update star by id
      tags:
      - stars
...
schemes:
- http
swagger: "2.0"